Trying to integrate blob and contouring
With this commit, I'm attempting to move the blob detection after the contouring. The contouring has been changed to now fill the contours, so that the blob detection works better. This doesn't work amazingly right now, so using the previous commit with increased blurring may be good enough for non-competition stuff